refactor: rename files to align with 'ng style guide'
