+++ app/src/app/components/home/classes/pdf-protect-mode.enum.ts
