+++ app/src/app/components/home/password-input/classes/password-input.ts
