+++ app/src/app/components/password-gen/classes/password-options.ts
